feat: seperate dialogs and modules for farm page
