+++ lib/pages/farm/components/MyFarmItem.dart
@@ -1,0 +1,187 @@
+import 'package:ax_dapp/pages/farm/dialogs/RewardClaimDialog.dart';
+import 'package:ax_dapp/pages/farm/dialogs/UnstakeDialog.dart';
+import 'package:get/get.dart';
+import 'package:flutter/material.dart';
+
+import 'package:ax_dapp/service/Controller/Farms/Farm.dart';
+import 'package:ax_dapp/pages/farm/components/SingleLogoFarmTitle.dart';
+import 'package:ax_dapp/pages/farm/components/DoubleLogoFarmTitle.dart';
+import 'package:ax_dapp/pages/farm/modules/PageTextStyle.dart';
+import 'package:ax_dapp/pages/farm/modules/BoxDecoration.dart';
+
+// First card of the my farms page is unique
+Widget MyFarmItem(BuildContext context, bool isWeb, Farm farm,
+    double listHeight, double layoutWidth) {
+  double minCardHeight = 450;
+  double maxCardHeight = 500;
+  double cardWidth = isWeb ? 500 : layoutWidth;
+  double cardHeight = listHeight * 0.7;
+  if (cardHeight < minCardHeight) cardHeight = minCardHeight;
+  if (cardHeight > maxCardHeight) cardHeight = maxCardHeight;
+
+  TextStyle txStyle = textStyle(Colors.grey[600]!, 14, false, false);
+
+  Widget farmTitleWidget;
+  if (farm.athlete == null)
+    farmTitleWidget = SingleLogoFarmTitle(context, isWeb, farm, cardWidth);
+  else
+    farmTitleWidget = DoubleLogoFarmTitle(context, isWeb, farm, cardWidth);
+
+  return Container(
+    margin: isWeb
+        ? EdgeInsets.symmetric(horizontal: 10)
+        : EdgeInsets.symmetric(vertical: 10),
+    padding: EdgeInsets.symmetric(horizontal: 20),
+    height: cardHeight,
+    width: cardWidth,
+    decoration: boxDecoration(
+        Color(0x80424242).withOpacity(0.25), 20, 1, Colors.grey[600]!),
+    child: Column(
+      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      children: <Widget>[
+        // Farm Title
+        farmTitleWidget,
+        //Upper information section
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: <Widget>[
+            Text(
+              "Current APR",
+              style: txStyle,
+            ),
+            Text("${farm.dAPR.toStringAsFixed(2)}%", style: txStyle)
+          ],
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: <Widget>[
+            Text("TVL", style: txStyle),
+            Text("\$${farm.dTVL.toStringAsFixed(2)}", style: txStyle)
+          ],
+        ),
+        //Divider line
+        Divider(
+          thickness: 0.35,
+          color: Colors.grey[400],
+        ),
+        //Bottom information section
+        Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: <Widget>[
+              Text("Your Position",
+                  style: textStyle(Colors.white, 20, false, false)),
+            ]),
+        //Show different information for AX item card and AX with APT card
+        if (farm.athlete == null) ...[
+          Container(
+            width: cardWidth,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Text(
+                  "Currently Staked",
+                  style: txStyle,
+                ),
+                Obx(() => Text(
+                    "${farm.dStaked.toStringAsFixed(2)} ${farm.strStakedSymbol}",
+                    style: txStyle))
+              ],
+            ),
+          ),
+          Container(
+            width: cardWidth,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Text(
+                  "Rewards Earned",
+                  style: txStyle,
+                ),
+                Obx(() => Text(
+                    "${farm.dRewards.toStringAsFixed(2)} ${farm.strRewardSymbol}",
+                    style: txStyle))
+              ],
+            ),
+          ),
+          Container(
+            width: cardWidth,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Text("Total AX available (Staked + Earned)", style: txStyle),
+                Obx(() => Text(
+                    "${(farm.dStaked.value + farm.dRewards.value).toStringAsFixed(2)} ${farm.strRewardSymbol}",
+                    style: txStyle))
+              ],
+            ),
+          ),
+        ] else ...[
+          Container(
+            width: cardWidth,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Text(
+                  "LP tokens provided",
+                  style: txStyle,
+                ),
+                Text("100 LP", style: txStyle)
+              ],
+            ),
+          ),
+          Container(
+            width: cardWidth,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Text("Rewards Earned", style: txStyle),
+                Obx(() => Text(
+                    "{farmController.rewardsEarned} ${farm.strRewardSymbol}",
+                    style: txStyle))
+              ],
+            ),
+          ),
+        ],
+        //Claim rewards and Unstake liquidity buttons
+        Container(
+          width: cardWidth,
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            children: <Widget>[
+              Container(
+                  //subtract padding for card's content for mobile
+                  width: isWeb ? 215 : (cardWidth / 2) - 25,
+                  height: 35,
+                  decoration: boxDecoration(
+                      Colors.amber[600]!, 100, 0, Colors.amber[600]!),
+                  child: TextButton(
+                      onPressed: () async => {
+                            await farm.claim(),
+                            showDialog(
+                                context: context,
+                                builder: (BuildContext context) =>
+                                    RewardClaimDialog(context))
+                          },
+                      child: Text("Claim Rewards",
+                          style: textStyle(Colors.black, 14, true, false)))),
+              Container(
+                  //width takes into account padding for card's content
+                  width: isWeb ? 240 : (cardWidth / 2) - 25,
+                  height: 35,
+                  decoration: boxDecoration(
+                      Colors.transparent, 100, 0, Colors.amber[600]!),
+                  child: TextButton(
+                      onPressed: () => showDialog(
+                          context: context,
+                          builder: (BuildContext context) =>
+                              UnstakeDialog(context, farm, cardWidth, isWeb)),
+                      child: Text("Unstake Liquidity",
+                          style:
+                              textStyle(Colors.amber[600]!, 14, true, false)))),
+            ],
+          ),
+        ),
+      ],
+    ),
+  );
+}
