+++ lib/pages/farm/modules/BoxDecoration.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+BoxDecoration boxDecoration(
+    Color col, double rad, double borWid, Color borCol) {
+  return BoxDecoration(
+      color: col,
+      borderRadius: BorderRadius.circular(rad),
+      border: Border.all(color: borCol, width: borWid));
+}
