+++ lib/pages/farm/modules/DialogTextStyle.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+TextStyle textStyle(Color color, double size, bool isBold) {
+  if (isBold)
+    return TextStyle(
+      color: color,
+      fontFamily: 'OpenSans',
+      fontSize: size,
+      fontWeight: FontWeight.w500,
+    );
+  else
+    return TextStyle(
+      color: color,
+      fontFamily: 'OpenSans',
+      fontSize: size,
+    );
+}
