+++ lib/service/Dialog.dart
@@ -3,7 +3,6 @@ import 'package:ax_dapp/pages/connectWallet/MobileLoginPage.dart';
 import 'package:ax_dapp/pages/scout/models/AthleteScoutModel.dart';
 import 'package:ax_dapp/service/ApproveButton.dart';
 import 'package:ax_dapp/service/Controller/Controller.dart';
-import 'package:ax_dapp/service/Controller/Farms/Farm.dart';
 import 'package:ax_dapp/service/Controller/Pool/PoolController.dart';
 import 'package:ax_dapp/service/Controller/Scout/LSPController.dart';
 import 'package:ax_dapp/service/Controller/WalletController.dart';
@@ -291,448 +290,6 @@ Dialog walletDialog(BuildContext context) {
         ],
       ),
     ),
-  );
-}
-
-Dialog depositDialog(
-    BuildContext context, Farm farm, double layoutWdt, bool isWeb) {
-  double _height = MediaQuery.of(context).size.height;
-  double wid = isWeb ? 390 : layoutWdt;
-  double hgt = _height < 455 ? _height : 450;
-  double dialogHorPadding = 30;
-
-  Farm selectedFarm = Farm.fromFarm(farm);
-  TextEditingController stakeAxInput = TextEditingController();
-  WalletController walletController = Get.find();
-  RxDouble totalStakedBalance = 0.0.obs;
-
-  return Dialog(
-    //remove inset padding to increase width of child widget
-    insetPadding: EdgeInsets.zero,
-    backgroundColor: Colors.transparent,
-    shape: RoundedRectangleBorder(
-      borderRadius: BorderRadius.circular(12.0),
-    ),
-    child: Container(
-      height: hgt,
-      width: wid,
-      padding: EdgeInsets.symmetric(vertical: 22, horizontal: dialogHorPadding),
-      decoration: boxDecoration(Colors.grey[900]!, 30, 0, Colors.black),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Text(
-                "Stake Liquidity",
-                style: textStyle(Colors.white, 20, false),
-              ),
-              Container(
-                  child: TextButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      child: Icon(
-                        Icons.close,
-                        size: 30,
-                        color: Colors.white,
-                      )))
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              //Amount Box
-              Container(
-                margin: const EdgeInsets.symmetric(vertical: 30),
-                // padding: const EdgeInsets.all(10),
-                // Amount box was overflowing by 30px after using dialogHorPadding
-                width: wid - dialogHorPadding - 30,
-                height: 55,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                  borderRadius: BorderRadius.circular(14.0),
-                  border: Border.all(
-                    color: Colors.grey[400]!,
-                    width: 0.5,
-                  ),
-                ),
-                //Amount box content
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: <Widget>[
-                    //Icon image
-                    Container(
-                      width: 35,
-                      height: 35,
-                      decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        image: DecorationImage(
-                          image: AssetImage("assets/images/x.jpg"),
-                        ),
-                      ),
-                    ),
-                    //Empty space between icon image and ticker
-                    Container(width: 15),
-                    Expanded(
-                      child: Text(
-                        "${selectedFarm.strStakedAlias.value.length > 0 ? selectedFarm.strStakedAlias : selectedFarm.strStakedSymbol}",
-                        style: textStyle(Colors.white, 15, false),
-                      ),
-                    ),
-                    //Max button
-                    Container(
-                      height: 28,
-                      width: 48,
-                      decoration: boxDecoration(
-                          Colors.transparent, 100, 0.5, Colors.grey[400]!),
-                      child: TextButton(
-                        onPressed: () {
-                          walletController
-                              .getTokenBalance(farm.strStakeTokenAddress)
-                              .then((tokenBalance) {
-                            stakeAxInput.text = tokenBalance;
-                            selectedFarm.dStakeBalance.value =
-                                double.parse(tokenBalance);
-                            totalStakedBalance.value =
-                                selectedFarm.dStaked.value +
-                                    selectedFarm.dStakeBalance.value;
-                            print(stakeAxInput);
-                          });
-                        },
-                        child: Text(
-                          "Max",
-                          style: textStyle(Colors.grey[400]!, 9, false),
-                        ),
-                      ),
-                    ),
-                    SizedBox(
-                      width: 80,
-                      child: TextFormField(
-                        controller: stakeAxInput,
-                        onChanged: (value) {
-                          selectedFarm.dStakeBalance.value =
-                              double.parse(value);
-                          totalStakedBalance.value =
-                              selectedFarm.dStaked.value +
-                                  selectedFarm.dStakeBalance.value;
-                        },
-                        style: textStyle(Colors.grey[400]!, 22, false),
-                        decoration: InputDecoration(
-                          hintText: '0.00',
-                          hintStyle: textStyle(Colors.grey[400]!, 22, false),
-                          contentPadding: const EdgeInsets.all(9),
-                          border: InputBorder.none,
-                        ),
-                        inputFormatters: [
-                          FilteringTextInputFormatter.allow(
-                              (RegExp(r'^(\d+)?\.?\d{0,6}'))),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                "Current ${selectedFarm.strStakedSymbol} Staked",
-                style: textStyle(Colors.grey[400]!, 14, false),
-              ),
-              Obx(() => Text(
-                    "${selectedFarm.dStaked.toString()} ${selectedFarm.strStakedSymbol}",
-                    style: textStyle(Colors.grey[400]!, 14, false),
-                  )),
-            ],
-          ),
-          Row(
-            children: [
-              Container(
-                padding: EdgeInsets.only(left: 55),
-                child: Text(
-                  "+",
-                  style: textStyle(Colors.grey[400]!, 14, false),
-                ),
-              )
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                "Funds Added",
-                style: textStyle(Colors.grey[400]!, 14, false),
-              ),
-              Obx(() => Text(
-                    "${selectedFarm.dStakeBalance.value.toString()} ${selectedFarm.strStakedSymbol}",
-                    style: textStyle(Colors.grey[400]!, 14, false),
-                  )),
-            ],
-          ),
-          Container(
-            margin: const EdgeInsets.symmetric(vertical: 10),
-            child: Divider(
-              thickness: 0.35,
-              color: Colors.grey[400],
-            ),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text("New Staked Balance"),
-              Obx(() => Text(
-                  "${totalStakedBalance.value.toString()} ${selectedFarm.strStakedSymbol}")),
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              ApproveButton(175, 45, 'Approve', selectedFarm.approve,
-                  selectedFarm.stake, transactionConfirmed)
-            ],
-          )
-        ],
-      ),
-    ),
-  );
-}
-
-// dynamic
-Dialog dualDepositDialog(BuildContext context, Farm selectedFarm,
-    String athlete, double layoutWdt, bool isWeb) {
-  double _height = MediaQuery.of(context).size.height;
-  double wid = isWeb ? 390 : layoutWdt;
-  double hgt = _height < 455 ? _height : 450;
-  double dialogHorPadding = 30;
-
-  TextEditingController stakeAxInput = TextEditingController();
-  WalletController walletController = Get.find();
-
-  return Dialog(
-    insetPadding: EdgeInsets.zero,
-    backgroundColor: Colors.transparent,
-    shape: RoundedRectangleBorder(
-      borderRadius: BorderRadius.circular(12.0),
-    ),
-    child: Container(
-        height: hgt,
-        width: wid,
-        padding: EdgeInsets.symmetric(horizontal: dialogHorPadding),
-        decoration: boxDecoration(Colors.grey[900]!, 30, 0, Colors.black),
-        child: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              Container(
-                  width: wid,
-                  margin: EdgeInsets.only(top: 25, bottom: 10),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: <Widget>[
-                      Container(
-                          child: Text(
-                        "Deposit Liquidity",
-                        style: textStyle(Colors.white, 20, false),
-                      )),
-                      Container(
-                          child: TextButton(
-                              onPressed: () {
-                                Navigator.pop(context);
-                              },
-                              child: Icon(
-                                Icons.close,
-                                size: 30,
-                                color: Colors.white,
-                              )))
-                    ],
-                  )),
-              Container(
-                  width: wid,
-                  margin: EdgeInsets.symmetric(vertical: 5),
-                  child: Text(
-                    "*Add liquidity to supply LP tokens to your wallet\nDeposit LP tokens to AX rewards",
-                    style: textStyle(Colors.grey[600]!, 11, false),
-                  )),
-              //Amount Box
-              Container(
-                width: wid,
-                height: 55,
-                decoration: boxDecoration(
-                    Colors.transparent, 14, 0.5, Colors.grey[400]!),
-                padding: EdgeInsets.symmetric(horizontal: 10),
-                margin: EdgeInsets.symmetric(vertical: 10),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: <Widget>[
-                    Container(width: 10),
-                    Container(
-                      width: 35,
-                      height: 35,
-                      decoration: BoxDecoration(
-                        shape: BoxShape.circle,
-                        image: DecorationImage(
-                          image: AssetImage("assets/images/x.jpg"),
-                        ),
-                      ),
-                    ),
-                    Container(width: 15),
-                    Expanded(
-                      child: Text(
-                        "AX",
-                        style: textStyle(Colors.white, 15, false),
-                      ),
-                    ),
-                    Container(
-                      height: 28,
-                      width: 48,
-                      decoration: boxDecoration(
-                          Colors.transparent, 100, 0.5, Colors.grey[400]!),
-                      child: TextButton(
-                        onPressed: () {
-                          walletController.getYourAxBalance().then((value) {
-                            stakeAxInput.text =
-                                walletController.yourBalance.value;
-                            print(stakeAxInput);
-                          });
-                        },
-                        child: Text(
-                          "Max",
-                          style: textStyle(Colors.grey[400]!, 9, false),
-                        ),
-                      ),
-                    ),
-                    SizedBox(
-                      width: 70,
-                      child: TextFormField(
-                        controller: stakeAxInput,
-                        onChanged: (value) {
-                          stakeAxInput.text = value;
-                        },
-                        style: textStyle(Colors.grey[400]!, 22, false),
-                        decoration: InputDecoration(
-                          hintText: '0.00',
-                          hintStyle: textStyle(Colors.grey[400]!, 22, false),
-                          contentPadding: const EdgeInsets.all(9),
-                          border: InputBorder.none,
-                        ),
-                        inputFormatters: [
-                          FilteringTextInputFormatter.allow(
-                              (RegExp(r'^(\d+)?\.?\d{0,6}'))),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              //Amount Box
-              Container(
-                  width: wid,
-                  height: 55,
-                  decoration: boxDecoration(
-                      Colors.transparent, 14, 0.5, Colors.grey[400]!),
-                  padding: EdgeInsets.symmetric(horizontal: 10),
-                  margin: EdgeInsets.symmetric(vertical: 10),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: <Widget>[
-                      Container(width: 10),
-                      Container(
-                        width: 35,
-                        height: 35,
-                        decoration: BoxDecoration(
-                          shape: BoxShape.circle,
-                          image: DecorationImage(
-                            scale: 0.5,
-                            image: AssetImage("assets/images/apt.png"),
-                          ),
-                        ),
-                      ),
-                      Container(width: 15),
-                      Expanded(
-                        child: Text(
-                          athlete + " APT",
-                          style: textStyle(Colors.white, 15, false),
-                        ),
-                      ),
-                      Container(
-                        height: 28,
-                        width: 48,
-                        decoration: boxDecoration(
-                            Colors.transparent, 100, 0.5, Colors.grey[400]!),
-                        child: TextButton(
-                          onPressed: () {},
-                          child: Text(
-                            "Max",
-                            style: textStyle(Colors.grey[400]!, 9, false),
-                          ),
-                        ),
-                      ),
-                      SizedBox(
-                        width: 70,
-                        child: TextFormField(
-                          onChanged: (value) {},
-                          style: textStyle(Colors.grey[400]!, 22, false),
-                          decoration: InputDecoration(
-                            hintText: '0.00',
-                            hintStyle: textStyle(Colors.grey[400]!, 22, false),
-                            contentPadding: const EdgeInsets.all(9),
-                            border: InputBorder.none,
-                          ),
-                          inputFormatters: [
-                            FilteringTextInputFormatter.allow(
-                                (RegExp(r'^(\d+)?\.?\d{0,6}'))),
-                          ],
-                        ),
-                      ),
-                    ],
-                  )),
-              Container(
-                  width: 175,
-                  height: 45,
-                  decoration: boxDecoration(
-                      Colors.transparent, 100, 1, Colors.amber[400]!),
-                  margin: EdgeInsets.only(top: 20, bottom: 10),
-                  child: TextButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                        showDialog(
-                            context: context,
-                            builder: (BuildContext context) =>
-                                transactionConfirmed(context));
-                      },
-                      child: Text("Add Liquidity",
-                          style: textStyle(Colors.amber[400]!, 20, true)))),
-              Container(
-                margin: EdgeInsets.symmetric(vertical: 10),
-                child: Text("LP Tokens: " + "0",
-                    style: textStyle(Colors.white, 18, true)),
-              ),
-              Container(
-                width: 175,
-                height: 40,
-                decoration: boxDecoration(Colors.grey, 100, 1, Colors.grey),
-                margin: EdgeInsets.symmetric(vertical: 10),
-                child: TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                    showDialog(
-                        context: context,
-                        builder: (BuildContext context) =>
-                            transactionConfirmed(context));
-                  },
-                  child:
-                      Text("Deposit", style: textStyle(Colors.black, 16, true)),
-                ),
-              )
-            ],
-          ),
-        )),
   );
 }
 
@@ -1328,94 +885,6 @@ Dialog confirmTransaction(
                           style: isWeb
                               ? textStyle(Colors.black, 16, false)
                               : textStyle(Colors.amber[500]!, 16, false),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ))));
-}
-
-// dynamic
-Dialog rewardsClaimed(BuildContext context) {
-  double _height = MediaQuery.of(context).size.height;
-  double _width = MediaQuery.of(context).size.width;
-  double wid = _width < 505 ? _width : 500;
-  double hgt = _height < 340 ? _height : 335;
-  double edge = 40;
-
-  return Dialog(
-      backgroundColor: Colors.transparent,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12.0),
-      ),
-      child: Container(
-          height: hgt,
-          width: wid,
-          decoration: boxDecoration(Colors.grey[900]!, 30, 0, Colors.black),
-          child: Center(
-              child: Container(
-            height: 275,
-            width: wid - edge,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Container(
-                    width: wid - edge,
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: <Widget>[
-                        Container(width: 5),
-                        Container(
-                          child: Text("Rewards Claimed",
-                              style: textStyle(Colors.white, 20, false)),
-                        ),
-                        Container(
-                          width: 40,
-                          child: IconButton(
-                            icon: const Icon(
-                              Icons.close,
-                              color: Colors.white,
-                            ),
-                            onPressed: () {
-                              Navigator.pop(context);
-                            },
-                          ),
-                        )
-                      ],
-                    )),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: <Widget>[
-                    Container(
-                      child: Icon(
-                        Icons.check_circle_outline,
-                        size: 150,
-                        color: Colors.amber[400],
-                      ),
-                    ),
-                  ],
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: <Widget>[
-                    Container(
-                      width: 275,
-                      height: 50,
-                      decoration: BoxDecoration(
-                        color: Colors.amber[400],
-                        borderRadius: BorderRadius.circular(100),
-                      ),
-                      child: TextButton(
-                        onPressed: () => Navigator.pop(context),
-                        child: Text(
-                          "View on Polygonscan",
-                          style: TextStyle(
-                            fontSize: 16,
-                            color: Colors.black,
-                          ),
                         ),
                       ),
                     ),
@@ -2026,230 +1495,6 @@ Dialog accountDialog(BuildContext context) {
             )
           ],
         )))),
-  );
-}
-
-// dynamic
-Dialog removeDialog(
-    BuildContext context, Farm farm, double layoutWdt, bool isWeb) {
-  double _height = MediaQuery.of(context).size.height;
-  double wid = isWeb ? 390 : layoutWdt;
-  double hgt = _height < 455 ? _height : 450;
-  double dialogHorPadding = 30;
-
-  Farm selectedFarm = Farm.fromFarm(farm);
-  RxDouble totalStakedBalance = 0.0.obs;
-  TextEditingController unStakeAxInput = TextEditingController();
-
-  return Dialog(
-    insetPadding: EdgeInsets.zero,
-    backgroundColor: Colors.transparent,
-    shape: RoundedRectangleBorder(
-      borderRadius: BorderRadius.circular(12.0),
-    ),
-    child: Container(
-      height: hgt,
-      width: wid,
-      padding: EdgeInsets.symmetric(vertical: 22, horizontal: dialogHorPadding),
-      decoration: boxDecoration(Colors.grey[900]!, 30, 0, Colors.black),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: <Widget>[
-              Text(
-                "Unstake Liquidity",
-                style: textStyle(Colors.white, 20, false),
-              ),
-              Container(
-                  child: TextButton(
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      child: Icon(
-                        Icons.close,
-                        size: 30,
-                        color: Colors.white,
-                      )))
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              //Amount Box
-              Container(
-                margin: const EdgeInsets.symmetric(vertical: 30),
-                width: wid - dialogHorPadding - 30,
-                height: 55,
-                decoration: BoxDecoration(
-                  color: Colors.transparent,
-                  borderRadius: BorderRadius.circular(14.0),
-                  border: Border.all(
-                    color: Colors.grey[400]!,
-                    width: 0.5,
-                  ),
-                ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: <Widget>[
-                    Container(
-                      height: 35,
-                      width: 35,
-                      decoration: BoxDecoration(
-                        image: DecorationImage(
-                          scale: 2.0,
-                          image: AssetImage('assets/images/x.jpg'),
-                        ),
-                        shape: BoxShape.circle,
-                      ),
-                    ),
-                    Container(width: 15),
-                    Expanded(
-                      child: Text(
-                        "${selectedFarm.strStakedAlias.value.length > 0 ? selectedFarm.strStakedAlias : selectedFarm.strStakedSymbol}",
-                        style: textStyle(Colors.white, 15, false),
-                      ),
-                    ),
-                    Container(
-                      height: 28,
-                      width: 48,
-                      decoration: boxDecoration(
-                          Colors.transparent, 100, 0.5, Colors.grey[400]!),
-                      child: TextButton(
-                        onPressed: () {
-                          unStakeAxInput.text =
-                              selectedFarm.dStaked.value.toString();
-                          selectedFarm.dUnStakeBalance.value =
-                              double.parse(unStakeAxInput.text);
-                          totalStakedBalance.value =
-                              selectedFarm.dStaked.value -
-                                  selectedFarm.dUnStakeBalance.value;
-                        },
-                        child: Text(
-                          "Max",
-                          style: textStyle(Colors.grey[400]!, 9, false),
-                        ),
-                      ),
-                    ),
-                    SizedBox(
-                      width: 70,
-                      child: TextFormField(
-                        controller: unStakeAxInput,
-                        onChanged: (value) {
-                          selectedFarm.dUnStakeBalance.value =
-                              double.parse(value);
-                          totalStakedBalance.value =
-                              selectedFarm.dStaked.value -
-                                  selectedFarm.dUnStakeBalance.value;
-                        },
-                        style: textStyle(Colors.grey[400]!, 22, false),
-                        decoration: InputDecoration(
-                          hintText: '0.00',
-                          hintStyle: textStyle(Colors.grey[400]!, 22, false),
-                          contentPadding: const EdgeInsets.all(9),
-                          border: InputBorder.none,
-                        ),
-                        inputFormatters: [
-                          FilteringTextInputFormatter.allow(
-                              (RegExp(r'^(\d+)?\.?\d{0,6}'))),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                "Current ${selectedFarm.strStakedSymbol} Staked",
-                style: textStyle(Colors.grey[400]!, 14, false),
-              ),
-              Obx(() => Text(
-                    "${selectedFarm.dStaked} ${selectedFarm.strStakedSymbol}",
-                    style: textStyle(Colors.grey[400]!, 14, false),
-                  )),
-            ],
-          ),
-          Row(
-            children: [
-              Container(
-                padding: EdgeInsets.only(left: 55),
-                child: Text(
-                  "-",
-                  style: textStyle(Colors.grey[400]!, 14, false),
-                ),
-              )
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                "Funds Removed",
-                style: textStyle(Colors.grey[400]!, 14, false),
-              ),
-              Obx(() => Text(
-                    "${selectedFarm.dUnStakeBalance} ${selectedFarm.strStakedSymbol}",
-                    style: textStyle(Colors.grey[400]!, 14, false),
-                  )),
-            ],
-          ),
-          Container(
-            margin: const EdgeInsets.symmetric(vertical: 10),
-            child: Divider(
-              thickness: 0.35,
-              color: Colors.grey[400],
-            ),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text("New Staked Balance"),
-              Obx(() =>
-                  Text("$totalStakedBalance ${selectedFarm.strStakedSymbol}")),
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              /* Container(
-                margin: EdgeInsets.only(top: 30.0, bottom: 10.0),
-                width: 175,
-                height: 45,
-                decoration: BoxDecoration(
-                  color: Colors.amber[400],
-                  borderRadius: BorderRadius.circular(100),
-                ),
-                child: TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                    showDialog(
-                        context: context,
-                        builder: (BuildContext context) =>
-                            removalConfimed(context));
-                  },
-                  child: const Text(
-                    "Confirm",
-                    style: TextStyle(
-                      fontSize: 16,
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ), */
-              //ApproveButton(175, 45, 'confirm', false, () => {}, () => {}),
-              ApproveButton(175, 45, 'Approve', testFunction,
-                  selectedFarm.unstake, removalConfirmed),
-            ],
-          )
-        ],
-      ),
-    ),
   );
 }
